# fix(query): the hashmap will cause oom