--- a/src/query/functions-v2/Cargo.toml
+++ b/src/query/functions-v2/Cargo.toml
@@ -30,6 +30,7 @@ criterion = "0.4"
 hex = "0.4.3"
 itertools = "0.10.5"
 match-template = "0.0.1"
+memchr = "2.5.0"
 num-traits = "0.2.15"
 once_cell = "1.15.0"
 ordered-float = { version = "3.1.0", features = [