I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Now `l_col like r_col` will generate a hashmap based on r_col, if r_col is huge, it will be oom kill.

sys environment:

> OS: Ubuntu 22.04.1 LTS (Jammy Jellyfish)
> CPU: Intel(R) Xeon(R) Platinum 8269CY CPU @ 2.50GHz
> Memory: 30G
> Disk: 16G / 197G

### Table Struct

```sql
-- databend
 CREATE TABLE `orders` (
  `o_orderkey` BIGINT,
  `o_custkey` BIGINT,
  `o_orderstatus` VARCHAR,
  `o_totalprice` DOUBLE,
  `o_orderdate` DATE,
  `o_orderpriority` VARCHAR,
  `o_clerk` VARCHAR,
  `o_shippriority` INT,
  `o_comment` VARCHAR
)
-- ck
SELECT version()

Query id: 73c60cca-5362-4765-b4fa-9e0a7a9cdd8b

┌─version()──┐
│ 22.10.2.11 │
└────────────┘

CREATE TABLE tpch.orders
(
    `o_orderkey` Int64,
    `o_custkey` Int32,
    `o_orderstatus` String,
    `o_totalprice` Decimal(15, 2),
    `o_orderdate` Date,
    `o_orderpriority` String,
    `o_clerk` String,
    `o_shippriority` Int32,
    `o_comment` String
)
ENGINE = MergeTree
ORDER BY o_orderkey
SETTINGS index_granularity = 8192

```

### Prepare Data

data from: https://github.com/leiysky/tpch-databend 

```shell
./tpch.sh 10.0
bash insert_data.sh
```

```sql
-- databend
mysql> select count() from orders ;
+----------+
| count()  |
+----------+
| 15000000 |
+----------+
1 row in set (0.01 sec)
Read 1 rows, 1.00 B in 0.001 sec., 796.38 rows/sec., 796.38 B/sec.


-- ck
SELECT count()
FROM orders

Query id: 173ca818-69f0-4935-8526-429cb321eac5

┌──count()─┐
│ 15000000 │
└──────────┘

1 row in set. Elapsed: 0.001 sec. 

```

### Result

test sql: `select count() from orders where o_clerk not  like o_comment;`

| before optimize | drop hash map | after optimize | ck |
|--|--|--|--|
| oom |114s|0.832 s|7.925s|

Closes #8615