@TCeason 2022-11-10
ali cloud 16C32G
all settings is default.

ck    databend
7.9s  114s

```sql
-- databend
mysql> select count() from orders where o_clerk not  like o_comment;
+----------+
| count()  |
+----------+
| 15000000 |
+----------+
1 row in set (1 min 53.77 sec)
Read 15000000 rows, 1.11 GiB in 113.761 sec., 131.86 thousand rows/sec., 10.00 MiB/sec.


-- ck


SELECT version()

Query id: 58b06879-7ca4-4854-96d1-d4dd02cc2540

┌─version()──┐
│ 22.10.2.11 │
└────────────┘


SELECT count()
FROM orders
WHERE o_clerk NOT LIKE o_comment

Query id: 0d1ff665-04e1-435d-bbb7-e21da85dff00

┌──count()─┐
│ 15000000 │
└──────────┘

1 row in set. Elapsed: 7.925 sec. Processed 15.00 million rows, 1.22 GB (1.89 million rows/s., 153.73 MB/s.)


```