@Xuanwo 2022-11-10
How does clickhouse handle this?