@Xuanwo 2022-11-10
BTW, like pattern is much simpler than regex. Maybe we don't need to convert like pattern into regex?