@Xuanwo 2022-11-10
> BTW, like pattern is much simpler than regex. Maybe we don't need to convert like pattern into regex?

I port a quick demo from [tikv](https://github.com/tikv/tikv/blob/master/components/tidb_query_expr/src/impl_like.rs)

The benches show that plain like is much faster than regex build (if we have to):

```shell
regex/build_regex       time:   [11.830 µs 11.857 µs 11.902 µs]
                        change: [-5.3625% -4.7534% -4.1822%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  8 (8.00%) low mild
  3 (3.00%) high severe
regex/build_like        time:   [3.1063 ns 3.1160 ns 3.1277 ns]
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe
```

<details>
<summary>The source code</summary>

```rust
fn bench_regex(c: &mut Criterion) {
    let mut group = c.benchmark_group("regex");

    group.bench_function("build_regex", |b| {
        b.iter(|| {
            let _ = regex::Regex::new("^Hello\\..World.*%$");
        })
    });
    group.bench_function("build_like", |b| {
        b.iter(|| {
            let _ = like(b"Hello, World", b"^Hello\\..World.*%$", &0);
        })
    });
}

fn decode_one(data: &[u8]) -> Option<(u8, usize)> {
    if data.is_empty() {
        None
    } else {
        Some((data[0], 1))
    }
}

fn sort_compare(a: &[u8], b: &[u8]) -> std::cmp::Ordering {
    a.cmp(b)
}

fn like(target: &[u8], pattern: &[u8], escape: &i64) -> Option<i64> {
    let escape = *escape as u32;
    // current search positions in pattern and target.
    let (mut px, mut tx) = (0, 0);
    // positions for backtrace.
    let (mut next_px, mut next_tx) = (0, 0);
    while px < pattern.len() || tx < target.len() {
        if let Some((c, mut poff)) = decode_one(&pattern[px..]) {
            let code: u32 = c.into();
            if code == '_' as u32 {
                if let Some((_, toff)) = decode_one(&target[tx..]) {
                    px += poff;
                    tx += toff;
                    continue;
                }
            } else if code == '%' as u32 {
                // update the backtrace point.
                next_px = px;
                px += poff;
                next_tx = tx;
                next_tx += if let Some((_, toff)) = decode_one(&target[tx..]) {
                    toff
                } else {
                    1
                };
                continue;
            } else {
                if code == escape && px + poff < pattern.len() {
                    px += poff;
                    poff = if let Some((_, off)) = decode_one(&pattern[px..]) {
                        off
                    } else {
                        break;
                    }
                }
                if let Some((_, toff)) = decode_one(&target[tx..]) {
                    if let std::cmp::Ordering::Equal =
                        sort_compare(&target[tx..tx + toff], &pattern[px..px + poff])
                    {
                        tx += toff;
                        px += poff;
                        continue;
                    }
                }
            }
        }
        // mismatch and backtrace to last %.
        if 0 < next_tx && next_tx <= target.len() {
            px = next_px;
            tx = next_tx;
            continue;
        }
        return Some(false as i64);
    }

    Some(true as i64)
}
```

</detials>