@BohuTANG 2022-11-10
This case is not a constant regex rule,  it's column-1 like column-2, is more difficult here.